hfp: move hfp_h2_framing into hfp_codec
